Fin du projet avec le tips , observe et numberformat

+++ Semaine1/TipsCalculator/app/src/main/java/ca/mikeyy/tipscalculator/domain/models/TipsCalculation.kt
@@ -4,7 +4,7 @@ import ca.mikeyy.tipscalculator.core.Constants
 
 data class TipsCalculation(private val subtotal : Double,private val percentageTip : Int) {
 
-    val tipsAmount = subtotal * (percentageTip / Constants.PERCENT)
+    val tipsAmount = subtotal * (percentageTip.toDouble() / Constants.PERCENT)
     val TPS = subtotal * Constants.pTPS
     val TVQ = subtotal * Constants.pTVQ
     val total = subtotal + tipsAmount + TPS + TVQ
